# backend/openapi.yaml
openapi: 3.0.3
info:
  title: MATRUVA Backend API
  description: |
    REST API for MATRUVA platform with JWT-based authentication, RBAC, and admin management.

    ## Authentication
    This API uses RS256 JWT tokens for authentication:
    - **Access Token**: Short-lived (15 min), used for API requests
    - **Refresh Token**: Long-lived (30 days), httpOnly cookie for token rotation

    ## Authorization
    Role-Based Access Control (RBAC) with three default roles:
    - `USER`: Regular users
    - `ADMIN`: System administrators
    - `SUPER_ADMIN`: Full system access

    ## Security
    - All admin endpoints require authentication
    - CSRF protection on token refresh via X-Auth-Refresh header
    - Refresh token rotation on every use
    - Session tracking with IP and User-Agent
  version: 1.0.0
  contact:
    name: MATRUVA Support
    email: support@matruva.com
  license:
    name: MIT
servers:
  - url: http://localhost:4000
    description: Development server
  - url: https://api.matruva.com
    description: Production server
...
paths:
...
  /v1/admin/keys:
    get:
      tags:
        - Admin - Keys
      summary: List JWT signing keys
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of keys
          content:
            application/json:
              schema:
                type: object
                properties:
                  keys:
                    type: array
                    items:
                      $ref: "#/components/schemas/KeyEntry"

    post:
      tags:
        - Admin - Keys
      summary: Upload new signing key
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - publicKey
                - use
              properties:
                publicKey:
                  type: string
                  description: PEM-encoded RSA public key
                privateKey:
                  type: string
                  description: PEM-encoded RSA private key (optional)
                use:
                  type: string
                  enum: [sig, enc]
                alg:
                  type: string
                  default: RS256
      responses:
        "201":
          description: Key uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  kid:
                    type: string
                  use:
                    type: string
                  active:
                    type: boolean
  /v1/admin/keys/{kid}:
    patch:
      tags:
        - Admin - Keys
      summary: Activate/deactivate key
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: kid
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - active
              properties:
                active:
                  type: boolean
      responses:
        "200":
          description: Key status updated

    delete:
      tags:
        - Admin - Keys
      summary: Delete key (soft delete)
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: kid
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Key deleted

# ============================================
# COMPONENTS
# ============================================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token in Authorization header

    CookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: Access token in httpOnly cookie

    RefreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
      description: Refresh token in httpOnly cookie (path restricted to /v1/auth)
...
  schemas:
...
    KeyEntry:
      type: object
      properties:
        kid:
          type: string
          description: Key ID
        use:
          type: string
          enum: [sig, enc]
        kty:
          type: string
          enum: [RSA]
        alg:
          type: string
          example: RS256
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        hasPrivateKey:
          type: boolean
...